Thanks to Paulo Ribeiro for catching this: if a page has no search document and you try to delete it, you get a MySQL error because of an empty WHERE IN clause. Fixed
git-svn-id: http://svn.apostrophenow.org/plugins/apostropheMysqlSearchPlugin/trunk@4094 c89fab08-2f9b-42bf-8176-bd3ad6bee547

+++ lib/aMysqlSearch.class.php
@@ -144,7 +144,11 @@ class aMysqlSearch extends aSearchService
     {
       $this->deleteUsages($document_id);
     }
-    $this->sql->query('DELETE FROM a_search_document WHERE id IN :document_ids', array('document_ids' => $document_ids));
+    // Careful, WHERE IN bombs on empty lists. Thanks to Paulo Ribeiro
+    if (count($document_ids))
+    {
+      $this->sql->query('DELETE FROM a_search_document WHERE id IN :document_ids', array('document_ids' => $document_ids));
+    }
   }
 
   /**
